+++ src/main/java/prismatic/shards/stellarity/registry/item/SatchelOfVoids.java
@@ -1,10 +1,10 @@
 package prismatic.shards.stellarity.registry.item;
 
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.context.UseOnContext;
 import org.jspecify.annotations.NonNull;
+import prismatic.shards.stellarity.registry.entity.SatchelSigil;
 
 public class SatchelOfVoids extends Item {
 	public SatchelOfVoids(Properties properties) {
@@ -19,12 +19,13 @@ public class SatchelOfVoids extends Item {
 	public @NonNull InteractionResult useOn(@NonNull UseOnContext useOnContext) {
 		var prev = super.useOn(useOnContext);
 		var player = useOnContext.getPlayer();
-		if (player == null || !player.level().isClientSide()) return prev;
+		var level = useOnContext.getLevel();
+		if (player == null || level.isClientSide()) return prev;
 
-		Component msg = Component.literal("Not implemented yet! Thanks for finding this :3");
-
-
-		player.sendOverlayMessage(msg);
+		var sigil = new SatchelSigil(level);
+		sigil.setPos(useOnContext.getClickedPos().relative(useOnContext.getClickedFace()).getBottomCenter());
+		level.addFreshEntity(sigil);
+		player.getCooldowns().addCooldown(useOnContext.getItemInHand(), 70 * 20);
 
 		return InteractionResult.SUCCESS;
 	}